requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
amadeus==8.0.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.34.11
botocore==1.34.11
cachetools==5.3.1
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cryptography==39.0.2
cssutils==2.9.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==4.0.0
django-environ==0.11.2
django-excel==0.0.10
django-filter==22.1
django-inlinecss==0.3.0
django-jet-reboot==1.3.3
django-multiselectfield==0.1.12
django-resized==1.0.2
django-rest-passwordreset==1.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
dynaconf==3.1.11
et-xmlfile==1.1.0
facebook-sdk==3.1.0
fonttools==4.49.0
frozenlist==1.4.1
future==0.18.3
google-api-core==2.11.1
google-api-python-client==1.4.1
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
humanfriendly==9.2
humanize==4.9.0
hvac==1.0.2
idna==3.4
inflection==0.5.1
install==1.3.5
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.5
knox==0.1.14
lml==0.1.0
loguru==0.6.0
magic-filter==1.0.12
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
Pillow==9.4.0
protobuf==4.23.3
psycopg2-binary==2.9.6
py-sr25519-bindings==0.2.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyhcl==0.4.4
PyJWT==2.6.0
pynliner==0.8.0
pyparsing==3.0.9
pyreadline==2.1
pyrsistent==0.19.3
pyTelegramBotAPI==4.16.1
python-dateutil==2.8.2
python-social-auth==0.3.6
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.3
telebot==0.0.5
Telethon==1.34.0
texttable==1.7.0
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
vonage==3.13.0
vonage-jwt==1.1.0
win32-setctime==1.1.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
